Remove v1 support
